[Refactor] model_dialect -> framework_dialect
